Move API from __init__.py to model.py
